fetch_url plan thwarted by the buffer
The original fetch_url implementation did not work as intended, due to
buffering of stdout when using shell pipes. Although workarounds exist
(script and stdbuf among them), both the full version and BusyBox
version of wget support redirecting the status output with -o. This
approach sidesteps the buffering issue but means we have to parse the
status data only when we actually care about it. Thus, a skeleton
download_progress script has been started in ui for this eventual
purpose.

It was necessary to drop curl support to implement this change.