Finished refactoring download backends.
Download and information extraction is now working for curl, wget, and
wget2. Testing reveals that neither wget nor wget2 is really a great
tool for use as a download backend, however. In particular, the wget
output wildly extrapolates transfer speeds over time periods that are
too short, making its current speed display useless. The authors removed
that speed display from wget2, along with any projection of the
remaining download time. Consequently, speeds and remaining times have
to be calculated in awk.

In reality, calculating the speed is probably going to have to be
something to be done manually for wget, if wget is to be kept as a
backend. I'm not confident that the text preprocessing code for wget2
will work in all cases, since wget2 uses terminal control sequences to
implement its progress bar, and --progress=dot is not implemented in the
version shipping on Slackware 15.0.

The TL;DR is that I might refactor one more time and just hard-code a
dependency on curl. Supporting only one backend would greatly simplify
the download code.