Began actual refactoring work on download_file.
The purpose of this script is going to be changed to download multiple
files off a single server, since that is the most efficient way to
handle HTTP downloads when several files need to be retrieved from the
same host.

Most of today's actual work involved rewriting the comments, which no
longer match the current code implementation.